// src/roblouie_tmd/src/tmd/primitive-type.enum.ts
export enum PrimitiveType {
  LINE_SOLID = "LINE_SOLID",
  LINE_GRADIENT = "LINE_GRADIENT",

  THREE_SIDED_FLAT_NO_TEXTURE_SOLID = "3_SIDED_FLAT_NO_TEXTURE_SOLID",
  THREE_SIDED_GOURAD_NO_TEXTURE_SOLID = "3_SIDED_GOURAD_NO_TEXTURE_SOLID",
  THREE_SIDED_FLAT_NO_TEXTURE_GRADIENT = "3_SIDED_FLAT_NO_TEXTURE_GRADIENT",
  THREE_SIDED_FLAT_TEXTURE = "3_SIDED_FLAT_TEXTURE",
  THREE_SIDED_GOURAD_TEXTURE = "3_SIDED_GOURAD_TEXTURE",
  THREE_SIDED_NO_LIGHT_NO_TEXTURE_SOLID = "3_SIDED_NO_LIGHT_NO_TEXTURE_SOLID",
  THREE_SIDED_NO_LIGHT_TEXTURE_SOLID = "3_SIDED_NO_LIGHT_TEXTURE_SOLID",

  FOUR_SIDED_GOURAD_NO_TEXTURE_SOLID = "4_SIDED_GOURAD_NO_TEXTURE_SOLID",
  FOUR_SIDED_FLAT_TEXTURE_NO_COLOR = "4_SIDED_FLAT_TEXTURE",
  FOUR_SIDED_GOURAD_TEXTURE = "4_SIDED_GOURAD_TEXTURE",
  FOUR_SIDED_NO_LIGHT_NO_TEXTURE_SOLID = "4_SIDED_NO_TEXTURE_SOLID",
  FOUR_SIDED_NO_LIGHT_TEXTURED_SOLID = "4_SIDED_TEXTURE_SOLID",
}